pyproject.toml
[project]
name = "notify-niche"
version = "0.1.0"
description = "RSS Collector and Niche Information Notification System"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "feedparser>=6.0.0",
    "requests>=2.28.0",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
